Move resolving workspaceEnv to provision package
Resolve workspaceEnv from attributes when we're creating a deployment
rather than when we're flattening the DevWorkspace to ensure all
containers (i.e. including project-clone) get workspaceEnv applied as
expected.
